feat: Rename `packages/common/` to `packages/engines/`, and move things around.
The `Store` was originally present in the `wasmer-common-python` crate,
but it's not longer required. Same for the `errors` module. Both have
moved back into the `wasmer-python` crate.

Now `wasmer-common-python` only contains types about the engines and
sibling types (`Target`, `Triple` etc.). So it's time to rename
`wasmer-common-python` to `wasmer-engines-python`.

Boom, no more “common” generic crate.
