--- packages/common/README.md
@@ -1,6 +1,0 @@
-# `wasmer_common`
-
-This crate is not compiled to a Python package. It represents common
-types that are shared between the `wasmer` and the `wasmer_compiler_*`
-packages mostly. Please read those packages' documentation to learn
-more.
