+++ packages/compiler-llvm/Cargo.toml
@@ -15,7 +15,7 @@ name = "wasmer_compiler_llvm"
 crate-type = ["cdylib"]
 
 [dependencies]
-wasmer-common-python = { path = "../common/" }
+wasmer-engines-python = { path = "../engines/" }
 wasmer-compiler-llvm = { git = "https://github.com/wasmerio/wasmer", branch = "master" }
 pyo3 = { version = "0.11", features = ["extension-module"] }
