+++ packages/engines/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "wasmer-common-python"
+name = "wasmer-engines-python"
 version = "0.5.0"
 authors = ["Wasmer Engineering Team <engineering@wasmer.io>"]
 edition = "2018"
-description = "Common types for the `wasmer-*-python` crates"
+description = "Engines and sibling types for the `wasmer-*-python` crates"
 readme = "README.md"
 repository = "https://github.com/wasmerio/wasmer-python"
 keywords = ["python", "extension", "webassembly"]
