+++ packages/engines/README.md
@@ -1,0 +1,6 @@
+# `wasmer_engines`
+
+This crate is not compiled to a Python package. It represents engines
+and sibling types that are shared between the `wasmer` and the
+`wasmer_compiler_*` packages mostly. Please read those packages'
+documentation to learn more.
