+++ packages/engines/src/lib.rs
@@ -1,0 +1,5 @@
+mod engines;
+mod target_lexicon;
+
+pub use crate::engines::{Native, OpaqueCompiler, JIT};
+pub use crate::target_lexicon::{CpuFeatures, Target, Triple};
